+++ bw.workflow.deleteconvictedreport/src/main/java/sa/gov/nic/bio/bw/workflow/deleteconvictedreport/controllers/EnterReportNumberPaneFxController.java
@@ -6,6 +6,7 @@ import javafx.scene.control.Button;
 import javafx.scene.control.ProgressIndicator;
 import javafx.scene.control.TextField;
 import sa.gov.nic.bio.bw.core.Context;
+import sa.gov.nic.bio.bw.core.controllers.DevicesRunnerGadgetPaneFxController;
 import sa.gov.nic.bio.bw.core.controllers.WizardStepFxControllerBase;
 import sa.gov.nic.bio.bw.core.utils.FxmlFile;
 import sa.gov.nic.bio.bw.core.utils.GuiUtils;
@@ -32,6 +33,14 @@ public class EnterReportNumberPaneFxController extends WizardStepFxControllerBas
 		
 		btnNext.disableProperty().bind(txtReportNumber.textProperty().isEmpty().or(txtReportNumber.disabledProperty()));
 		txtReportNumber.requestFocus();
+		
+		DevicesRunnerGadgetPaneFxController deviceManagerGadgetPaneController =
+												Context.getCoreFxController().getDeviceManagerGadgetPaneController();
+		
+		if(!deviceManagerGadgetPaneController.isDevicesRunnerRunning())
+		{
+			deviceManagerGadgetPaneController.runAndConnectDevicesRunner();
+		}
 	}
 	
 	@Override