+++ bw.workflow.editconvictedreport/src/main/java/module-info.java
@@ -8,6 +8,6 @@ module bw.workflow.editconvictedreport
 	opens sa.gov.nic.bio.bw.workflow.editconvictedreport.bundles;
 	opens sa.gov.nic.bio.bw.workflow.editconvictedreport to bw.core;
 	opens sa.gov.nic.bio.bw.workflow.editconvictedreport.fxml to bw.core;
-	//opens sa.gov.nic.bio.bw.workflow.editconvictedreport.tasks to bw.core;
+	opens sa.gov.nic.bio.bw.workflow.editconvictedreport.tasks to bw.core;
 	opens sa.gov.nic.bio.bw.workflow.editconvictedreport.controllers to javafx.fxml, bw.core;
 }