+++ RELEASE-NOTES.md
@@ -15,6 +15,7 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 - ENH: Unify the transition overlay for menu transition and wizard transition.
 - ENH: Upgrade controlsfx library from 9.0.0 to 11.0.0.
+- NEW: Add Edit-Convicted-Report workflow.
 
 ---